backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.3
drf-yasg==1.21.7
Pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.7
gunicorn==21.2.0
openpyxl==3.1.2
